+++ FreeMat2/libs/libCore/Misc.cpp
@@ -2558,8 +2558,8 @@ namespace FreeMat {
     int Cm, Cn, Cm_offset, Cn_offset;
     Cm = X.getDimensionLength(0);
     Cn = X.getDimensionLength(1);
-    Cm_offset = floor((double)((Y.getDimensionLength(0)-1)/2));
-    Cn_offset = floor((double)((Y.getDimensionLength(1)-1)/2));
+    Cm_offset = (int) floor((double)((Y.getDimensionLength(0)-1)/2));
+    Cn_offset = (int) floor((double)((Y.getDimensionLength(1)-1)/2));
     return singleArrayVector(Conv2FunctionDispatch(X,Y,Cm,Cn,Cm_offset,Cn_offset));
   }
